Renamed pagination for search results
